backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Vadlamani Uday Kumar <kumar.princeofpersia.uday@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
django-ninja = "^1.3.0"
django-ninja-extra = "^0.21.8"
django-ninja-jwt = "^5.3.5"
pydantic = { extras = ["dotenv"], version = "^2.10.4" }
aioredis = "^2.0.1"
psycopg2 = "^2.9.10"
flake8 = "^7.1.1"
isort = "^5.13.2"
black = "^24.10.0"
autopep8 = "^2.3.1"
mypy = "^1.14.0"
pytest = "^8.3.4"
uvicorn = { extras = ["standard"], version = "^0.34.0" }

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "C90"]


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
